feat: add support for multiple tags and labels
We are introducing a new feature that will allow our customers to limit
scans not only by a single tag or a single label, but now they will be
able to pass multiple tags and labels.

We are deprecating old attributes that were of type `String` in favor of
new attributes of type `List(string)` and `Map<key, value>`.

Moved Attributes:
* `limit_by_tag` -> `limit_by_tags`
* `limit_by_label` -> `limit_by_labels`
* `limit_by_repos` -> `limit_by_repositories`

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>