docs: improve documentation for org account_mappings
Current documentation for Organization Level integration of Cloud
Accounts is not clear, this change is clarifying the need to define a
new Lacework provider with the argument `organization` set to `true` to
access the organization level data set.

It also adds a clear example of how to migrate an existing AWS
CloudTrail integration from an account to the organization level.

Closes https://github.com/lacework/terraform-provider-lacework/issues/142

JIRA https://lacework.atlassian.net/browse/ALLY-611

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>