refactor(lacework_query): remove query_id from Schema
The motivation of this change it to avoid making our users set the query
id twice, it makes no sense to allow them to configure a different query
id, specially not since the platform enforces both, the query id and the
query text to match the same id.

We are removing the `query_id` field from the Schema of the
`lacework_query` resource and instead, we are automatically guessing the
query id from the actual query.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>