chore(RAIN-98039): add the optional `queryLanguage` enum value where used
This change adds the `queryLanguage` enumeration to the lacework_query and lacework_policy_compliance resource types.

For a query, it is an optional part of the create submission. It is also an optional part of query payload response from the result of creating, updating, or retrieving the definition of a query.

For a policy, it is an optional part of the response from creating, updating or retrieving the definition of a policy.